# src/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Gestion UH API service for centralized user administration",
        "title": "Gestion UH API",
        "contact": {
            "email": "nodo@uh.cu"
        },
        "version": "1.0"
    },
    "host": "identity.sic.uh.cu",
    "paths": {
...
        "/email/quota": {
            "get": {
                "description": "Get email buzon quota",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email"
                ],
                "summary": "Get email buzon quota",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.EmailQuotaResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/proxy/quota": {
            "get": {
                "description": "Get Internet Quota Consumption",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy"
                ],
                "summary": "Get proxy quota",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ProxyQuotaResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/securityquestions": {
            "get": {
                "description": "Get available security questions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Questions"
                ],
                "summary": "Get security questions.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SecurityQuestions"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/user/changepassword": {
            "post": {
                "description": "Change user password stored in LDAP server.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Change user password.",
                "parameters": [
                    {
                        "description": "Old and new password",
                        "name": "newPassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ChangePasswordModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/user/getVPN": {
            "get": {
                "description": "Get VPN of given user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get VPN of given user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CI of user to get VPN",
                        "name": "ci",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/user/me": {
            "get": {
                "description": "Get User Data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userData"
                ],
                "summary": "Get User Data.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.UserDataModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/user/search": {
            "get": {
                "description": "Search user by CI",
                "summary": "Search user by CI",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Text with the query data",
                        "name": "text",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/user/securityquestions": {
            "get": {
                "description": "Get user stored security questions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Questions"
                ],
                "summary": "Get user security questions.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ci",
                        "name": "ci",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "email",
                        "name": "ci",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SecurityQuestions"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.ChangePasswordModel": {
            "type": "object",
            "required": [
                "newPassword",
                "oldPassword"
            ],
            "properties": {
                "newPassword": {
                    "type": "string"
                },
                "oldPassword": {
                    "type": "string"
                }
            }
        },
...
        "api.EmailQuotaResult": {
            "type": "object",
            "required": [
                "consumed",
                "quota"
            ],
            "properties": {
                "consumed": {
                    "type": "integer"
                },
                "quota": {
                    "type": "integer"
                }
            }
        },
        "api.Error": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "api.ProxyQuotaResult": {
            "type": "object",
            "required": [
                "bonus",
                "consumed",
                "quota"
            ],
            "properties": {
                "bonus": {
                    "type": "integer"
                },
                "consumed": {
                    "type": "integer"
                },
                "quota": {
                    "type": "integer"
                }
            }
        },
...
        "api.SecurityQuestions": {
            "type": "object",
            "required": [
                "questions"
            ],
            "properties": {
                "questions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "api.UserDataModel": {
            "type": "object",
            "required": [
                "activeYears",
                "careerName",
                "ci",
                "curseType",
                "email",
                "hasCloud",
                "hasEmail",
                "hasInternet",
                "hasVPN",
                "hasVc",
                "militancy",
                "name",
                "objectClass",
                "position",
                "salary",
                "scientificDegree",
                "serviceTime",
                "subArea",
                "subCategory",
                "vacations"
            ],
            "properties": {
                "activeYears": {
                    "type": "string"
                },
                "careerName": {
                    "type": "string"
                },
                "ci": {
                    "type": "string"
                },
                "curseType": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "hasCloud": {
                    "type": "boolean"
                },
                "hasEmail": {
                    "type": "boolean"
                },
                "hasInternet": {
                    "type": "boolean"
                },
                "hasVPN": {
                    "type": "boolean"
                },
                "hasVc": {
                    "type": "boolean"
                },
                "militancy": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "objectClass": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "salary": {
                    "type": "string"
                },
                "scientificDegree": {
                    "type": "string"
                },
                "serviceTime": {
                    "type": "string"
                },
                "subArea": {
                    "type": "string"
                },
                "subCategory": {
                    "type": "string"
                },
                "vacations": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}